restructuring permissions
